+++ domain.js
@@ -4,13 +4,20 @@
 'use strict';
 
 (function () {
-	var oldNodePath = '',
-		platform = process.platform,
-		NODE_PATH = process.env.NODE_PATH;
+	var fs					= require('fs'),
+		path				= require('path'),
+		Checker				= require('jscs'),
+		jscsConfig			= require('jscs/lib/cli-config'),
+		findup				= require('findup'),
+		domainName			= 'globexdesigns.brackets-jscs',
+		configFiles			= ['.jscsrc', '.jscs.json'],
+		oldNodePath			= '',
+		platform			= process.platform,
+		NODE_PATH			= process.env.NODE_PATH;
 
-	if (NODE_PATH) {
-		oldNodePath = NODE_PATH + (platform === 'win32' ? ';' : ':');
-	}
+	// Setup NODE_PATH configuration
+
+	if (NODE_PATH) oldNodePath = NODE_PATH + (platform === 'win32' ? ';' : ':');
 
 	if (platform === 'win32') {
 		NODE_PATH = oldNodePath + process.env.APPDATA + '\\npm\\node_modules';
@@ -22,33 +29,72 @@
 
 	require('module').Module._initPaths();
 
-	var path = require('path'),
-		Checker = require('jscs'),
-		jscsConfig = require('jscs/lib/cli-config'),
-		cli = new Checker(),
-		currentProjectRoot = null,
-		domainName = 'evnaverniouk.brackets-jscs';
+	var _findConfig = function (fullPath, callback) {
+		findup(fullPath, function (dir, cb) {
+			var found = false,
+				configpath;
 
-	function _setProjectRoot(projectRoot) {
+			for (var i = 0, l = configFiles.length; i < l; i++) {
+				configpath = path.join(dir, configFiles[i]);
+				if (fs.existsSync(configpath)) {
+					found = true;
+					break;
+				}
+			}
 
-		cli = new Checker();
-		console.log("NODE PATH", NODE_PATH)
-		
-		var config = jscsConfig.load(path.join(projectRoot, '.jscsrc'))
-		
-		cli.configure(config);
-	}
+			cb(found);
+		}, function (err, dir) {
+			if (err) return callback(null);
+			callback(dir);
+		});
+	};
 
-	function lintFile(fullPath, projectRoot, callback) {
-		if (projectRoot !== currentProjectRoot) {
-			_setProjectRoot(projectRoot);
-			currentProjectRoot = projectRoot;
+	var _getConfigFile = function (fullPath) {
+		var config = null,
+			configpath;
+
+		for (var i = 0, l = configFiles.length; i < l; i++) {
+			configpath = path.join(fullPath, configFiles[i]);
+			if (fs.existsSync(configpath)) {
+				config = jscsConfig.load(configpath);
+				break;
+			}
 		}
 
-		return cli.checkPath(fullPath).then(function (response) {
-			callback(response[0]._errorList);
+		return config || {};
+	};
+
+
+	var _setConfig = function (cli, configPath) {
+		var config = _getConfigFile(configPath);
+
+		// Load JSCS configuration
+		cli.registerDefaultRules();
+		cli.configure(config);
+	};
+
+
+	var lintFile = function (fullPath, projectRoot, callback) {
+		var cli = new Checker();
+
+		return _findConfig(path.dirname(fullPath), function (result) {
+			// If no config file found - let users know
+			if (!result) {
+				return callback([{
+					message: "Unable to find a JSCS configuration file."
+				}]);
+			}
+
+			// Set config file
+			_setConfig(cli, result);
+
+			// Run JSCS checker
+			cli.checkPath(fullPath).then(function (response) {
+				callback(response[0]._errorList);
+			});
 		});
-	}
+	};
+
 
 	exports.init = function (domainManager) {
 		if (!domainManager.hasDomain(domainName)) {
@@ -71,9 +117,9 @@
 				name: 'projectRoot',
 				type: 'string'
 			}], [{
-                name: "errors",
-                type: "array"
-            }]
+				name: "errors",
+				type: "array"
+			}]
 		);
 	};
 }());